mock: move to internal/backend
